+++ ksqldb-rest-app/src/main/java/io/confluent/ksql/rest/server/KsqlServerEndpoints.java
@@ -18,6 +18,9 @@ package io.confluent.ksql.rest.server;
 import static io.netty.handler.codec.http.HttpResponseStatus.NOT_FOUND;
 
 import io.confluent.ksql.api.auth.ApiSecurityContext;
+import io.confluent.ksql.api.impl.InsertsStreamEndpoint;
+import io.confluent.ksql.api.impl.KsqlSecurityContextProvider;
+import io.confluent.ksql.api.impl.QueryEndpoint;
 import io.confluent.ksql.api.server.InsertResult;
 import io.confluent.ksql.api.server.InsertsStreamSubscriber;
 import io.confluent.ksql.api.spi.Endpoints;
